requirements.txt
Flask==2.3.2
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
beautifulsoup4==4.12.2
demjson3==3.0.6
python-dotenv==1.0.0
